common/gpio: Set GPIO flags after setting alternate function
On STM32, gpio_set_alternate_function() called with GPIO_ALT_FUNC_NONE
argument causes the pin to be set to input. In gpio_reset() function,
we are setting flags first and then we are changing pin function to
GPIO. As a result, using gpio_reset() on STM32 boards causes pin to be
configured as input, regardles of gpio.inc file.

To fix this, we should always change pin functionality to GPIO before
setting GPIO flags and/or state. This approach seems to be natural also.

BUG=b:170432597
BRANCH=none
TEST=Flash EC on some STM32 based board (eg. dartmonkey).
     Use gpio_reset() function on some GPIO pin which is configured as
     GPIO_OUT_LOW/GPIO_OUT_HIGH in gpio.inc file (eg. FP_RST_ODL on
     dartmonkey).
     Use `gpioset` and `gpioget` command to change state of the GPIO
     pin, and verify it works.

Signed-off-by: Patryk Duda <pdk@semihalf.com>
Change-Id: I0a6c2ee8d1c1cf33b14bd54bb432c63b820427be