zephyr: herobrine_npcx9: Initial image which uses NPCX9
This CL copies the Zephyr board/trogdor to board/herobrine_npcx9.
Modify the chip config to npcx9m3f and some related configs (simply
rename) and dts (update the UART property). The board/herobrine_npcx9
and other NPCX9 boards, like board/brya will be merged to a single
board. But the merge needs more work to remove the not-common configs
and dts to the project directories. Leave it as a future work.

This CL also copies the project trogdor/herobrine_npcx7 to
herobrine/herobrine_npcx9. Remove the dead files power.c and
hibernate.c, in the EC-OS directory. Remove the Zephyr 2.5 support
and board.cmake (needed only for 2.5).

BRANCH=None
BUG=b:192253134
TEST=Built the herobrine_npcx9 image successfully.

Change-Id: I9641768ee978920f6a8677f13ae14e0a26ad35f4
Signed-off-by: Wai-Hong Tam <waihong@google.com>