USB MUX: Generalize mux ACK
Currently, only the virtual mux driver uses the mux ACK feature, but the
actual wait for the host command ACK is a part of the usb_mux general
code.  Generalize this mux ACK wait so it's available if needed in the
future for more muxes.

Additionally, moving this wait out of the mux set will allow us to lock
the muxes intelligently between tasks, without keeping the muxes locked
during the inactive ACK wait.

BRANCH=None
BUG=b:172222942,b:186777984
TEST=tast typec.Mode*.manual on voxel

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I61a043425a482cc6f3170548c888d91ec20c2a82