Remove a few bytes from debug message
This CL removes a few bytes from a debug message.
This is because the Mushu board almost has no space left in FLASH.
The next CL in the chain add some new logic that requires some extra
bytes, but will fail on Mushu due to lack on space.

BUG=None
BRANCH=None
TEST=make -j BOARD=mushu (in particular, mushu passes)
     make -j buildall
     Those were compiled in locally. Builder might need a few less (or
     more) bytes.

Signed-off-by: Ricardo Quesada <ricardoq@google.com>
Change-Id: I2a0879483c5900432d2173afe47e20fe34c4d474