cbi: Add CONFIG_EEPROM_CBI_WP
This commit adds the config option, CONFIG_EEPROM_CBI_WP.  It is to be
defined when the EC directly controls the CBI EEPROM WP signal.  The EC
will set the WP according to the result of `system_is_locked()`.  Note
that once the WP is set, the EC must be reset via EC_RST_ODL in order
for the WP to become unset.  This is enabled by the accompanying
hardware.

BUG=b:181769483
BRANCH=None
TEST=`make -j buildall`

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: If490594ab4dd24af98119b01299215b997913b66