chgstv2/sustainer: Stop AC current when discharging
Currently, the battery sustainer discharges the battery using
CHARGE_CONTROL_IDLE, which uses the AC current and stops the
current from the battery.

This CL makes the sustainer discharge the battery using
CHARGE_CONTROL_IDLE, which stops the AC current and uses the
battery power to drive the system.

BUG=b:188457962
BRANCH=None
TEST=buildall

Change-Id: I4af31eff488bc9cfa627f84994b685488c3c9061
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>