temp_sensor: Support reading Millikelvin
This change adds an extra output parameter for millikelvin to the
temp_sensor read method. When millikelvin resolution is not supported,
temp_mk_ptr is set to -1. This is done so developers do not erroneously
assume a sensor supports millikelvin resolution.

temp_sensor_read_mk should be used when millikelvin resolution is
required. It will return EC_ERROR_UNIMPLEMENTED if millikelvin
resolution is not supported by a given sensor.

BUG=b:192391025
BRANCH=None
TEST=Build all and pass unit tests

Signed-off-by: Rob Barnes <robbarnes@google.com>
Change-Id: I53019d162814ad393bede22d7230587c2a79f4db