test: Fix printf formatting in usb_typec_drp_acc_trysrc.c
Fix a couple of minor formatting errors where constants with LL or ULL
suffixes were being printed. On x86_64 printing these as %l is fine
because longs are 64-bits. On i686 this comes out as an error because
long is 32-bits. Use the correct ll specifier for these, and in one case
ensure it comes out as a 64-bit value by adding LL.

BUG=b:179062230
BRANCH=none
TEST=make -j runhosttests on i686 and x86_64

Signed-off-by: Evan Green <evgreen@chromium.org>
Change-Id: Ic59c9647aa70f58e497b9615122ca4c202504365