test: Fix sbs_charging_v2's charge_control() cmd version
charge_control() was sending EC_CMD_CHARGE_CONTROL version 2, but
leaving params.cmd uninitialized. Apparently we were just getting lucky.
I did not get as lucky when compiling in 32-bit mode. This test is
clearly sending the version 1 form of the command, so make the version
number match the intent so the EC doesn't fail the command.

BUG=b:179062230
BRANCH=none
TEST=make -j runhosttests BOARD=host on x86_64 and i686

Signed-off-by: Evan Green <evgreen@chromium.org>
Change-Id: Ie440ff825e1125bb50b4ac2477ba983e00ef303e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3059229
Reviewed-by: Diana Z <dzigterman@chromium.org>