test: Make more and smaller allocations in shmalloc
The shmalloc test is randomly allocating and freeing shared memory of
size (rand % total_available). One of the tests assertions is that it
was able to get at least 5 allocations this way. This is essentially a
game of blackjack where we happened to get a bunch of low cards. We're
not necessarily a shoe-in with this algorithm, and when I tried
compiling for 32-bits I found I did not get so lucky. The test was
failing for me because it could not happen to get 5 allocations, it
spent most of its time bouncing between 2 and 4.

There are two options: either decrease 5 to 4, or don't let rand take
quite so much per allocation. More smaller allocations for a malloc-like
function seemed like a better test than fewer large ones. So let's
decrease the max allocation to total_available / 2.

BUG=b:179062230
BRANCH=none
TEST=make -j runhosttests BOARD=host and x86_64 and i686.

Signed-off-by: Evan Green <evgreen@chromium.org>
Change-Id: Ifb16533ee6fa02ce2613c32b1462b78ef0c6cd21