test: Ensure time move forward in usb_pd_int
test_storm_not_triggered_for_32bit_overflow wanted to check to ensure
the storm detection worked across the 32-bit boundary, so it forced the
current time to 0xff000000. It assumed however that this timestamp was
in the future. If it's in the past (say for instance someone else has
bumped time forward) then the storm tracking is broken and this test
fails. Since it is not the goal of this test to see how the storm
tracker behaves when time moves backwards, fix up the test to ensure
it's got the high bit set but is also in the future.

BUG=b:179062230
BRANCH=none
TEST=make -j runhosttests

Signed-off-by: Evan Green <evgreen@chromium.org>
Change-Id: Iac4c68e2bf46649a01477dbd1dabfcf0d95b1371