biod: Add fptool.py that will replace flash_fp_mcu
This is similar to gsctool, but this tool would only be in the
test image.

This comes from the effort to move away from shell script.
See http://go/deshell.

The idea is that you start with a simple wrapper for the given shell
script. Over time, features are transplanted or added-new to the python
wrapper, until the shell script is no longer needed.

BRANCH=none
BUG=b:172020576
TEST=./util/fptool.py
     ./util/fptool.py --help
     ./util/fptool.py flash --help
     ./util/fptool.py flash
     # Error about flash_fp_mcu not existing
TEST=scp util/fptool.py $DUT:/usr/local/bin
     ssh $DUT fptool.py flash /root/doesnotexist
     ssh $DUT fptool.py flash
     ssh $DUT fptool.py flash /opt/google/biod/fw/$(cros_config /fingerprint board)*.bin

Change-Id: I8774bda4a057ee09f70bf474ce3ba2fa0bbcf92d
Signed-off-by: Craig Hesling <hesling@chromium.org>