stm32g4: Fix RW write protect flash page count
This CL fixes a bug related to protection flags. The last page index
was being compared to the total number of flash pages resulting in an
off by one error. This would result in the flag
EC_FLASH_PROTECT_ALL_AT_BOOT not being set and that would in turn
cause the flash_pre_init routine to try again and force a reset. This
resulted in an continuous boot looping.

BRANCH=quiche
BUG=b:194991631
TEST=Verifed that following power on reset,
EC_FLASH_PROTECT_ALL_AT_BOOT gets set if the EC_WP_L line is asserted
and baklava boots up normally.

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I42c93ad3f8c5373f18a8a85ae4f547182805760f