USB MUX: Gate ACK event send on ACK waiting
Currently, the EC doens't always wait for an ACK from the virtual mux
but it will still always send the AP ACK onto the PD task.  This means
we have some risk of a previous ACK being present in the task events
already when we begin a new wait.  Remove this risk by only sending the
event to the task when there is a task waiting.

BRANCH=None
BUG=b:186777984
TEST=tast typec.Mode*.manual tests on voxel

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I1e4cf96af838c0cbe4ef549337b304679b59d641