lid_angle: Create a common callback for lid angle changes
Each board defines its own callback lid_angle_peripheral_enable().
The implementation is very similar. Create a common implementation
and reduce the duplicated code.

This CL removes the board callbacks which are identifical to the
common callback. If it is slightly different, keep it and add
the __override tag.

The check of TEST_BUILD is unnecessary as the board callback is not
linked in the test build.

BRANCH=None
BUG=b:194922043
TEST=Build all the images.

Change-Id: I73d381730f35b80eff69399cdfc5fb54f839aee0
Signed-off-by: Wai-Hong Tam <waihong@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3069175
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
(cherry picked from commit 2eb53a39662b905c65015bafb5f73ab2324d369c)