driver: lis2dw12: Set FIFO vector number properly
We have been lucky the accelerometer was always in first position,
as the FIFO filling routine always set the sensor number to 0.
Fix for hayato.

BUG=b:192649615
TEST=On hayato, using
iioservice_simpleclient --channels="accel_x accel_y accel_z"
--device_id=0 --frequency=10 --samples=10
Check the sensor data is constant. (device_id 0 is the base
accelerometer), we get data for that sensor only.
Checking with device_id set to 2 (the lid accelerometer),
that we get samples for that sensor.
BRANCH=asurada

Signed-off-by: Gwendal Grignou <gwendal@google.com>
Change-Id: I96ea5f696c38b8c54aed7537f93eba70a647dc53