zephyr: sensors: Add a property to detect alternative sensors
Some boards may have various types of sensors, but it is not defined in
the CBI SSFC section. It is required to detect the sensors via I2C
transactions.

CrosEC has board-specific code to detect the sensors, but for Zephyr
add "detect-func" property to the alternative sensors node. It points
a function to detect the defined sensor. Replace the alt sensor in
the "motion_sensors" array after the detection.

BUG=b:188498050
BRANCH=none
TEST=Verify the motion sensors with 'accelinfo' command

Signed-off-by: Dawid Niedzwiecki <dn@semihalf.com>
Change-Id: Idf178f763d53edc6c2cafd4baecc4e43efacd73f