TCPMv2: Fix ordering of Vbus application in Attached.SRC entry
Per the TCPCI spec figures, we should be setting the CC lines and
enabling Auto Discharge Disconnect before attempting to source Vbus.
Doing these things out of order may cause Vbus sourcing to appear to
fail when a TCPC is waiting for ADD to set before enabling Vbus.

BRANCH=None
BUG=b:195966013
TEST=on guybrush, verify both C0 and C1 successfully detect that they
are sourcing Vbus with an unpowered Apple 3-in-1 dongle

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Icb0bd221a7d8eb069b74327c27894e5d92e95329