ANX7451: Always reset mux after power on
Since the ANX7451 is unpowered in Z1, return an error for mux sets
during this time.  This will let the usb_mux.c code know that any set it
attempted to perform was unsuccessful.  Otherwise, it will be assumed
that setting mux mode of None succeeded, when we'll actually power up to
default USB mode.

Additionally, have the board clear out any cached mux state when we hit
G3, as the mux will no longer be powered and will need to be re-set.

BRANCH=None
BUG=b:195045790
TEST=on guybrush, power board in various scenarios and ensure the retimer
is consistently powered off when nothing is plugged in

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Ic98be3bc15e8f0affeda25d7d7170b89e3aecf10