USB MUX: Introduce flag to clear mux state in G3
Some muxes may need to be re-set after they've power cycled, and many
muxes are tied to the S5 power rails.  Introduce a flag which can be
used to indicate a mux needs to re-run init and is no longer in LPM
after the G3 state has been hit.

BRANCH=None
BUG=b:195045790
TEST=make -j buildall

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I8bd4184dbea629edf106dbee32f811620ebda0dd