servo_v4p1: Manage USB2/USB3 terminations based on data role
Previously servo_v4p1 was running with USB2 and USB3 paths always on.
This was completely ignoring data role of servo. Especially, we
shouldn't apply device terminations on USB2/USB3 lines once servo is
acting as DFP. While in this mode the only allowed configuration is to
establish FASTBOOT, that is direct connection between DUT and HOST.

Make use of the CONFIG_USBC_SS_MUX_UFP_ONLY, since USB3 muxer on servo
should be enabled only when servo is UFP.

BUG=b:137887386,b:182419010
BRANCH=main
TEST=Perform "cc snkdts" and "cc srcdts" sequence couple of time and
     verify presence of all expected USB2 & USB3 devices on the DUT
     side.
TEST=Another test may be to unplug and re-plug CHG connector to servo
     and verify presence of all expected USB2 & USB3 devices on the DUT.

Signed-off-by: Jan Dabros <jsd@semihalf.com>
Change-Id: I54a637b67cf62f20658c8c4522a8d8ccc011d863