honeybuns: fix check for fw_config from CBI
This CL fixes a logic error where CBI FW_CONFIG field was being
checked. In addition, the CBI config option to bypass the write
protection check was added as the CBI is only being used to store the
dock_mf preference and doesn't need to be write protected.

BUG=b:164157329
BRANCH=quiche
TEST=Verified on Gingerbread that the status LED color matches with
the dock's MF preference. In addition, verified that following a power
cycle, the MF preference is read properly from the CBI.

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: Ie3d5b8d4af73b1d09b3f58f1a990b7c77cb4a6fe