CoachZ: Disable IRQ from CTN730 in RO
When PCHG isn't enabled in RO, a keyboard is disabled until another
USB device is plugged.

This patch restores PCHG in RO but disables it by assigning an
invalid value to the IRQ signal (GPIO_IH_COUNT). A PCHG task
will be started but will not run because there will be no event.

BUG=b:193223400, b:173235954
BRANCH=trogdor
TEST=On CoachZ, stylus charges in normal mode and doesn't charge in
     recovery mode.

Change-Id: Icb5702474ac5ef74119455ab4879447cb7a963e4
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>