driblee/raa489000: Control ASGATE state
There is 0.1V leakage from Charger to Vbus
when a charger is removed in S5.
Control ASGATE state to prevent the leakage.

This commit uses the charger side registers to
control the ASGATE when selecting our active charge port.
This is done in addition to the existing implementation
which uses the TCPCI registers to control ASGATE.

BUG=b:196923271
BRANCH=keeby
TEST=make BOARD=driblee

Signed-off-by: Matt_Wang <Matt_Wang@compal.corp-partner.google.com>
Change-Id: I9c43941fd1795e28456a959d0e46f8268e8ac071