cr50: replace direct calls to EC OS from FIPS module with callbacks
In order to implement self-integrity test for FIPS module we need to
make sure binary code of module in image doesn't change from build to
build. To do that we already place FIPS module as constant address.
However, any call to functions outside the module creates a relocation
which is changing depending on location of that external function in the
image. To prevent that we either need to bring these functions in the
module like it was done with memcpy() and some others or replace their
invocations with callbacks. Task & Memory management functions are hard
to bring in the module, so replace few invocations with indirect calls
using vtable. This way invocation code will remain the same.

1. Identify and minimize dependency on EC OS - remove few asserts
   and cprintfs.
2. Remove checking privilege level in TRNG init - we know that it is
   high by the order of initialization in board_init() and that our
   RO doesn't drop permissions. Correct initialization of TRNG is
   important for certification, so we can't just assume it may be
   initialized improperly.
3. Added vtable with EC OS functions, initialization of FIPS module
   vtable in board_init().
4. Switched to using vtable instead of direct calls. Note, we continue
   to use EC OS with CRYPTO_TEST=1 to reduce vtable size and image
   size.

BUG=b:138578318
TEST=make BOARD=cr50; tests

Signed-off-by: Vadim Sukhomlinov <sukhomlinov@google.com>
Change-Id: Ibd7bd2353fc4e7e5886f9bfef96b36dc64ff2359