zephyr: zmake: add re-exec logic
Add some logic to re-exec zmake when running inside of a chroot and we
find the source code available (to reduce the number of times we have
to update the chroot).

Note: we also discussed adding a chroot-upgrade-required warning
logic.  This is actually a separate idea so will be handled in a
separate CL.

BUG=b:197636145
BRANCH=none
TEST=unit tests

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I26ecbb9a575f22c8667a1928e4bd5836f6fd4fe1