stt: Add startup delay after S0
STT interface is not ready for several seconds after S0. Add a delay so
writes are not attempted for several seconds. Attempting a write early
will not result in a critical error, but it will cause error messages to
be printed to the log which will cause confusion.

BUG=b:197745639
TEST=No STT errors after booting with non-serial AP
BRANCH=None

Change-Id: Id0c926ce896cdb4d3096746b3e7cc99db83bdd2d
Signed-off-by: Rob Barnes <robbarnes@google.com>