CBI: Add keybox field
The keybox for Widevine will be stored in the CBI EEPROM on some
systems.  Due to its large size and the fact that it is unique per
system, it is only being set through the ectool utility.  The EC
command-line doesn't have a sufficient buffer to support a full key, and
the cbi-util tool is meant to create a general image for many systems to
be flashed to.

BRANCH=None
BUG=b:193870273
TEST=on guybrush, verify set from EC console fails.  Verify that ectool
is able to set valid byte strings to the field, and rejects invalid
strings such as "fish".  Verify ectool retrieves a byte array that looks
just like the input string given, ex. "01234567" returns bytes "01 23 45
67"

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Ibcdd4158c70d5a03eb8f1cf18ad9c1a5681d02e2