tcpm: Mitigate EC RX circular buffer overflow
There are two FIFO buffers for USB-PD packet reception. One is located
in  TCPC, another was added in EC layer in crrev/c/1185727 . There are
scenarios that may lead to overflow.

In these scenarios, the EC RX circular buffer overflows without clearing
the TCPC ALERT# signal. This in turn results in a tight-loop in the TCPC
interrupt handler, leading to port "suspend".

This CL is a reduced hotfix to mitigate this scenario. Further CL is
needed to perform proper TCPCI spec-compliant overflow handling.

BRANCH=none
BUG=b:192382194
TEST=1. Connect CB282K via DP to G5 dock on old FW
     2. Connect G5 dock to AKEMI on TCPMv1
     3. Allow MST VDM:Attention communication to occur
     4. Observe EC console prints via SuzyQ

Change-Id: Ic579a3763fc595b8250d4ca78c3942675450295a
Signed-off-by: Nathan K <nkolluru@google.com>