cr50: fix make buildall builds
When both https://crrev.com/c/3125994 and https://crrev.com/c/3119223
landed it resulted in broken host build as function attribute
optimize("") is not supported by clang. This is happened due as both CLs
are relatively independent by nature, but the first one added one of the
sources to host build.

BUG=none
TEST=make buildall -j

Signed-off-by: Vadim Sukhomlinov <sukhomlinov@google.com>
Change-Id: If1b8425cea418ae42179bef46f6f5998ded3dc1f